+++ game_patch/multi/server_internal.h
@@ -26,6 +26,25 @@ struct HitSoundsConfig
     int rate_limit = 10;
 };
 
+struct WeaponStayExemptionConfig
+{
+    bool enabled = false;
+    bool rocket_launcher = false;
+    bool heavy_machine_gun = false;
+    bool sniper_rifle = false;
+    bool assault_rifle = false;
+    bool machine_pistol = false;
+    bool shotgun = false;
+    bool scope_assault_rifle = false;
+    bool grenade = false;
+    bool remote_charge = false;
+    bool handgun = false;
+    bool flamethrower = false;
+    bool riot_stick = false;
+    bool riot_shield = false;
+    bool rail_gun = false;
+};
+
 struct ServerAdditionalConfig
 {
     VoteConfig vote_kick;
@@ -39,6 +58,7 @@ struct ServerAdditionalConfig
     std::optional<float> spawn_armor;
     int ctf_flag_return_time_ms = 25000;
     HitSoundsConfig hit_sounds;
+    WeaponStayExemptionConfig weapon_stay_exemptions;
     std::map<std::string, std::string> item_replacements;
     std::string default_player_weapon;
     std::optional<int> default_player_weapon_ammo;